refactor(st-09081): modularize monitoring alerts
