+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-07-07
-**Active Story:** ST-09081 - Modularize Monitoring Alert Manager and Tests (Ready)
+**Last Updated:** 2026-07-08
+**Active Story:** ST-09081 - Modularize Monitoring Alert Manager and Tests (In Progress)
 
 ---
