+++ planning/kanban-queue.md
@@ -1,11 +1,11 @@
 # Kanban Queue: AgentForge
 
-**Last Updated:** 2026-07-07
+**Last Updated:** 2026-07-08
 
 ## Queue Status Summary
 
-- **Ready:** 1 story
-- **In Progress:** 0 stories
+- **Ready:** 0 stories
+- **In Progress:** 1 story
 - **In Review:** 0 stories
 - **Blocked:** 0 stories
 - **Backlog:** 0 stories
@@ -14,14 +14,14 @@
 
 ## Ready
 
-- `ST-09081` - Modularize Monitoring Alert Manager and Tests
-  - Depends on: `ST-09016` (merged 2026-03-26)
+None currently.
 
 ---
 
 ## In Progress
 
-None currently.
+- `ST-09081` - Modularize Monitoring Alert Manager and Tests
+  - Depends on: `ST-09016` (merged 2026-03-26)
 
 ---
